% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpha}
\alias{alpha}
\alias{alpha_ipsl_cm5a_lr}
\alias{alpha_gfdl_esm2m}
\alias{alpha_miroc_esm_chem}
\alias{alpha_noresm1_m}
\title{Dirichlet parameters for the ISIMIP models}
\format{List with 4 elements
\describe{
\item{tas}{Matrix [12 x ngrid] of alpha parameters for temperature for each grid
cell.  Non-land grid cells will have \code{NA} values.}
\item{pr}{Matrix [12 x ngrid] of alpha parameters for precipitation  for each
grid cell.  Non-land grid cells will have \code{NA} values.}
\item{coord}{Data frame of grid cell coordinates.}
\item{time}{Data frame of month names and numbers.}
}}
\usage{
alpha_ipsl_cm5a_lr

alpha_gfdl_esm2m

alpha_miroc_esm_chem

alpha_noresm1_m
}
\description{
These datasets contain the the fits of the Dirichlet distribution alpha
parameters for temperature and precipitation for the four ISIMIP models
supported by this package.  These structures are suitable for passing as the
\code{alpha}
argument to \code{\link{monthly_downscaling}}.
}
\keyword{datasets}
