% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_downscaling.R
\name{check_names}
\alias{check_names}
\title{Check an object for required names}
\usage{
check_names(list, req_names, list_name = NULL)
}
\arguments{
\item{list}{an object, such as list or data frame that needs certain names}

\item{list_name}{an optional string that will be incorporated into the error message, default is NULL}

\item{req_name}{a vector of the required names in the list}
}
\value{
an error message if the list is missing a required name
}
\description{
Check an object for required names
}
\keyword{internal}
