% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_downscaling.R
\name{downscaling_component_api}
\alias{downscaling_component_api}
\title{Wrapper function for running downscaling from cassandra component}
\usage{
downscaling_component_api(frac, fld_data, coords, time, var)
}
\arguments{
\item{frac}{Name of monthly fraction dataset to use.}

\item{fld_data}{List of matrices of annual field data to downscale to monthly.}

\item{coords}{Coordinate matrix for the field data, such as the one
produced by \code{\link[fldgen]{coord_array}}.}

\item{time}{Vector of years for the input field matrices.}

\item{var}{String indicating the variable being downscaled.  Either 'tas' or
'pr'.}
}
\value{
List of matrices of downscaled data.  Grid cells are in rows, months
in columns (This is the opposite of the convention used everywhere else in
this package.)
}
\description{
Convert inputs received from the cassandra component to the format expected
by \code{\link{monthly_downscaling}} and pass the converted arguments.
}
\details{
The conversions performed are
\describe{
\item{frac}{Dataset is looked up by name.}
\item{fld_data}{The downscaling function wants a single matrix.  This
function accepts a list of matrices and calls the downscaling function on
each one sequentially.}
\item{coords}{The matrix of coordinate data is converted to a data
frame, and the index column is added.}
\item{time}{No conversion.}
\item{var}{No conversion.}
}
}
\keyword{internal}
