% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_downscaling.R
\name{harmonize_coordinates}
\alias{harmonize_coordinates}
\title{Harmonize the fraction and field grid cells by coordinates}
\usage{
harmonize_coordinates(frac, frac_coordinates, fld_coordinates, var)
}
\arguments{
\item{frac}{the downscaling fraction list, should be available as package data}

\item{frac_coordinates}{the data frame of the fraction grid cell coordinates should be available from frac}

\item{fld_coordinates}{the data frame of the field grid cell coordinates}

\item{var}{the variable, tas or pr, to process}
}
\value{
the fraction matrix to use in monthly downscaling arranged so that the fraction grid cells match the order of the field grid cells
}
\description{
Harmonize the fraction and field grid cells by coordinates
}
