% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_downscaling.R
\name{monthly_downscaling}
\alias{monthly_downscaling}
\title{Downscale annual gridded data to monthly gridded data and convert units}
\usage{
monthly_downscaling(frac, fld, fld_coordinates, fld_time, var)
}
\arguments{
\item{frac}{the matrix of the monthly fractions to use in downscaling, created by \code{harmonize_coordinates}}

\item{fld}{the full grid data to downscale}

\item{fld_coordinates}{the grid cell coordinates for the field matrix}

\item{fld_time}{a vector of the field time values}

\item{var}{the string of the variable to process}
}
\value{
a list containing a 2d data array, a data frame of grid cell coordinates, and a unit string
}
\description{
Downscale annual gridded data to monthly gridded data and convert units
}
