% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{pr_conversion}
\alias{pr_conversion}
\title{Convert from monthly downscaled preciptation from kg/m2*s to mm/month}
\usage{
pr_conversion(input)
}
\arguments{
\item{input}{a 2d array of monhtly downscaled preciptation data in kg/m2*s to convert to mm/month, row.names of input must correspond to a time formatted as YYYYMM}
}
\value{
a matrix of monthly precipitation data in mm/month
}
\description{
Convert from monthly downscaled preciptation from kg/m2*s to mm/month
}
