% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_downscaling.R
\name{reindex_grid}
\alias{reindex_grid}
\title{Re-index the grid so that coordinates match latitude and longitude}
\usage{
reindex_grid(frac, frac_coordinates, fld_coordinates, var)
}
\arguments{
\item{frac}{the downscaling fraction list, internal pacakge data}

\item{frac_coordinates}{the data frame of the fraction grid cell coordinates should be available from frac}

\item{fld_coordinates}{the data frame of the field grid cell coordinates}

\item{var}{the name of the variable to process, a string of tas or pr}
}
\value{
the fraction matrix to use in monthly downscaling arranged so that the fraction grid cells match the order of the field grid cells
}
\description{
\code{reindex_grid}  returns the fraction matrix to use in \code{monthly_downscaling} during the monthly downscaling process.
}
\details{
This function ensures that that the monthly fractions used in the downscaling and the data being downscaled have the same
latitude and longitude coordinate system. This is important when working with ISIMIP and CMIP5 files since ISMIP has NA
values that may be removed during an earlier step which would change the grid cell indexing of the data.
}
\keyword{internal}
