% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_downscaling.R
\name{monthly_downscaling}
\alias{monthly_downscaling}
\title{Downscale annual data to monthly data.}
\usage{
monthly_downscaling(alpha, fld_data, fld_coordinates, fld_time, var)
}
\arguments{
\item{alpha}{Matrix[ngrid, 12] of the alpha parameters for the Dirichlet
distribution.  These values are included as package data.  For more details
on how they are calculated see data-raw.}

\item{fld_data}{the 2d array of ntime x ngrid of the data to be downscaled, this object should be a flattened 3-D array, where latitude
the most rapidly varying index for the individual time slices.}

\item{fld_coordinates}{a data frame of the grid cells coordinates for the field matrix.}

\item{fld_time}{a vector of the field time values.}

\item{var}{the string of the variable to process.}
}
\value{
a list containing the monthly downsaceld 2d array and a
}
\description{
\code{monthly_downscaling} Use average monthly fractions of annual tas or pr ISMIP/CMIP5 data to downscale annual data to montly values.
}
