% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{pr_conversion}
\alias{pr_conversion}
\title{Convert from kg/m2*s to mm/month}
\usage{
pr_conversion(input)
}
\arguments{
\item{input}{a matrix of the monthly downscaled pr data}
}
\value{
a matrix of pr data in mm/month
}
\description{
Convert from kg/m2*s to mm/month
}
\keyword{internal}
