% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{tas_conversion}
\alias{tas_conversion}
\title{Convert tas from K to C}
\usage{
tas_conversion(input)
}
\arguments{
\item{input}{a 2d array of temperature data in degree K to convert to C}
}
\value{
a matrix of the monthly downscaled temperature in degrees C
}
\description{
Convert tas from K to C
}
\keyword{internal}
